--- a/src/dotnet/APIView/APIViewIntegrationTests/TestsBaseFixture.cs
+++ b/src/dotnet/APIView/APIViewIntegrationTests/TestsBaseFixture.cs
@@ -18,8 +18,6 @@ using APIView.Identity;
 using APIViewWeb.Managers;
 using APIViewWeb.Hubs;
 using Microsoft.AspNetCore.SignalR;
-using APIViewWeb.Managers.Interfaces;
-using Microsoft.CodeAnalysis.Host;
 using Microsoft.Extensions.Options;
 using Microsoft.ApplicationInsights;
 
@@ -30,6 +28,8 @@ namespace APIViewIntegrationTests
         private readonly CosmosClient _cosmosClient;
         private readonly BlobContainerClient _blobCodeFileContainerClient;
         private readonly BlobContainerClient _blobOriginalContainerClient;
+        private IConfigurationRoot _config;
+        private readonly string _cosmosDBname = "ManagerTestsDB";
         
         public PackageNameManager PackageNameManager { get; private set; }
         public ReviewManager ReviewManager { get; private set; }
@@ -45,13 +45,15 @@ namespace APIViewIntegrationTests
 
         public TestsBaseFixture()
         {
-            var config = new ConfigurationBuilder()
+            _config = new ConfigurationBuilder()
                .AddEnvironmentVariables(prefix: "APIVIEW_")
                .AddUserSecrets(typeof(TestsBaseFixture).Assembly)
                .Build();
 
+            _config["CosmosDBName"] = _cosmosDBname;
+
             var services = new ServiceCollection();
-            services.AddSingleton<IConfiguration>(config);
+            services.AddSingleton<IConfiguration>(_config);
             services.AddMemoryCache();
             services.AddSingleton<PackageNameManager>();
             services.AddSingleton<LanguageService, JsonLanguageService>();
@@ -72,24 +74,25 @@ namespace APIViewIntegrationTests
             PackageNameManager = serviceProvider.GetService<PackageNameManager>();
             User = TestUser.GetTestuser();
 
-            _cosmosClient = new CosmosClient(config["Cosmos:ConnectionString"]);
-            var dataBaseResponse = _cosmosClient.CreateDatabaseIfNotExistsAsync("APIViewV2").Result;
+            _cosmosClient = new CosmosClient(_config["Cosmos:ConnectionString"]);
+            var dataBaseResponse = _cosmosClient.CreateDatabaseIfNotExistsAsync(_config["CosmosDBName"]).Result;
             dataBaseResponse.Database.CreateContainerIfNotExistsAsync("Reviews", "/id").Wait();
             dataBaseResponse.Database.CreateContainerIfNotExistsAsync("APIRevisions", "/ReviewId").Wait();
             dataBaseResponse.Database.CreateContainerIfNotExistsAsync("Comments", "/ReviewId").Wait();
             dataBaseResponse.Database.CreateContainerIfNotExistsAsync("Profiles", "/id").Wait();
-            ReviewRepository = new CosmosReviewRepository(config, _cosmosClient);
-            APIRevisionRepository = new CosmosAPIRevisionsRepository(config, _cosmosClient);
-            CommentRepository = new CosmosCommentsRepository(config, _cosmosClient);
-            var cosmosUserProfileRepository = new CosmosUserProfileRepository(config, _cosmosClient);
+            dataBaseResponse.Database.CreateContainerIfNotExistsAsync("PullRe", "/id").Wait();
+            ReviewRepository = new CosmosReviewRepository(_config, _cosmosClient);
+            APIRevisionRepository = new CosmosAPIRevisionsRepository(_config, _cosmosClient);
+            CommentRepository = new CosmosCommentsRepository(_config, _cosmosClient);
+            var cosmosUserProfileRepository = new CosmosUserProfileRepository(_config, _cosmosClient);
 
-            _blobCodeFileContainerClient = new BlobContainerClient(config["Blob:ConnectionString"], "codefiles");
-            _blobOriginalContainerClient = new BlobContainerClient(config["Blob:ConnectionString"], "originals");
+            _blobCodeFileContainerClient = new BlobContainerClient(_config["Blob:ConnectionString"], "codefiles");
+            _blobOriginalContainerClient = new BlobContainerClient(_config["Blob:ConnectionString"], "originals");
             _ = _blobCodeFileContainerClient.CreateIfNotExistsAsync(PublicAccessType.BlobContainer);
             _ = _blobOriginalContainerClient.CreateIfNotExistsAsync(PublicAccessType.BlobContainer);
 
-            BlobCodeFileRepository = new BlobCodeFileRepository(config, memoryCache);
-            var blobOriginalsRepository = new BlobOriginalsRepository(config);
+            BlobCodeFileRepository = new BlobCodeFileRepository(_config, memoryCache);
+            var blobOriginalsRepository = new BlobOriginalsRepository(_config);
 
             var authorizationServiceMoq = new Mock<IAuthorizationService>();
             authorizationServiceMoq.Setup(_ => _.AuthorizeAsync(It.IsAny<ClaimsPrincipal>(), It.IsAny<Object>(), It.IsAny<IEnumerable<IAuthorizationRequirement>>()))
@@ -97,7 +100,7 @@ namespace APIViewIntegrationTests
 
             var telemetryClient = new Mock<TelemetryClient>();
 
-            var notificationManager = new NotificationManager(config, ReviewRepository, APIRevisionRepository, cosmosUserProfileRepository, telemetryClient.Object);
+            var notificationManager = new NotificationManager(_config, ReviewRepository, APIRevisionRepository, cosmosUserProfileRepository, telemetryClient.Object);
 
             var devopsArtifactRepositoryMoq = new Mock<IDevopsArtifactRepository>();
             devopsArtifactRepositoryMoq.Setup(_ => _.DownloadPackageArtifact(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<string>(), It.IsAny<string>(), It.IsAny<string>(), It.IsAny<string>()))
@@ -131,12 +134,12 @@ namespace APIViewIntegrationTests
                 commentsRepository: CommentRepository, languageServices: languageService, signalRHubContext: signalRHubContextMoq.Object,
                 telemetryClient: telemetryClient.Object);
 
-            TestDataPath = config["TestPkgPath"];
+            TestDataPath = _config["TestPkgPath"];
         }
 
         public void Dispose()
         {
-            _cosmosClient.GetDatabase("APIView").DeleteAsync().Wait();
+            _cosmosClient.GetDatabase(_cosmosDBname).DeleteAsync().Wait();
             _cosmosClient.Dispose();
 
             _blobCodeFileContainerClient.DeleteIfExists();