--- a/src/dotnet/APIView/APIViewWeb/Repositories/CosmosReviewRepository.cs
+++ b/src/dotnet/APIView/APIViewWeb/Repositories/CosmosReviewRepository.cs
@@ -6,6 +6,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using APIViewWeb.Helpers;
 using APIViewWeb.LeanModels;
 using APIViewWeb.Repositories;
 using Microsoft.Azure.Cosmos;
@@ -21,7 +22,7 @@ namespace APIViewWeb
 
         public CosmosReviewRepository(IConfiguration configuration, CosmosClient cosmosClient)
         {
-            _reviewsContainer = cosmosClient.GetContainer("APIViewV2", "Reviews");
+            _reviewsContainer = cosmosClient.GetContainer(configuration["CosmosDBName"], "Reviews");
             _legacyReviewsContainer = cosmosClient.GetContainer("APIView", "Reviews");
         }
 
@@ -57,6 +58,26 @@ namespace APIViewWeb
                 .WithParameter("@language", language)
                 .WithParameter("@isClosed", isClosed);
 
+            var itemQueryIterator = _reviewsContainer.GetItemQueryIterator<ReviewListItemModel>(queryDefinition);
+            var reviews = new List<ReviewListItemModel>();
+            while (itemQueryIterator.HasMoreResults)
+            {
+                var result = await itemQueryIterator.ReadNextAsync();
+                reviews.AddRange(result.Resource);
+            }
+            return reviews;
+        }
+
+        public async Task<List<ReviewListItemModel>> GetReviewsAsync(List<string> reviewIds, bool? isClosed = null)
+        {
+            var reviewIdsAsQueryStr = CosmosQueryHelpers.ArrayToQueryString(reviewIds);
+            var queryStringBuilder = new StringBuilder($"SELECT * FROM Reviews r WHERE r.id IN {reviewIdsAsQueryStr}");
+            if (isClosed != null)
+            {
+                queryStringBuilder.Append(" AND r.IsClosed = @isClosed");
+            }
+            var queryDefinition = new QueryDefinition(queryStringBuilder.ToString())
+               .WithParameter("@isClosed", isClosed);
             var itemQueryIterator = _reviewsContainer.GetItemQueryIterator<ReviewListItemModel>(queryDefinition);
             var reviews = new List<ReviewListItemModel>();
             while (itemQueryIterator.HasMoreResults)