- Improve CosmosPullRequestsRepository (performance bottleneck)
- Add HTML Minification

http://localhost:5000/Assemblies/Review/09de885a7c58444190fa9c9ae894d36d/632d733925b849c69c9624fdafb81d82?diffRevisionId=e70aca01c7db4226b8ec10c6e0d6a033&doc=False&diffOnly=False
Before
2.78 s server wait time
11.81 s content download time
14.59 s Total

After
3.82 s server wait time
1.35 s content download time
5.17 s Total